feat: Add BigQuery schema export and fix webapp
This commit introduces a script to export the BigQuery schema to a CSV file. It also includes several fixes for the webapp, including resolving broken links and fixing the Tote bet placement logic. A .gitignore file has been added to the project to exclude virtual environments and other unnecessary files from version control.